+++ sleepecg/utils.py
@@ -5,6 +5,7 @@
 """Utility functions."""
 
 
+import datetime
 import warnings
 from typing import Callable, Iterable, List, TypeVar
 
@@ -73,3 +74,22 @@ def _parallel(
     return Parallel(n_jobs=n_jobs)(
         delayed(function)(x, *args, **kwargs) for x in iterable
     )
+
+
+def _time_to_sec(time: datetime.time):
+    """
+    Convert a `datetime.time` to seconds.
+
+    `00:00:00` corresponds to `0` and `23:59:59` to `86399`.
+
+    Parameters
+    ----------
+    time : datetime.time
+        A `datetime.time` to convert.
+
+    Returns
+    -------
+    int
+        Time of day in seconds.
+    """
+    return time.hour * 3600 + time.minute * 60 + time.second